Create a discory page and move function from home to discovery
