+++ frontend/src/pages/Home.jsx
@@ -4,42 +4,6 @@ import EventCard from '../components/EventCard';
 import PageLayout from '../components/PageLayout';
 
 function Home() {
-  const [events, setEvents] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    const fetchEvents = async () => {
-      try {
-        const response = await api.get('/events/events');
-        setEvents(response.data);
-      } catch (err) {
-        setError(err);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchEvents();
-  }, []);
-
-  if (loading) {
-    return <div>Loading...</div>;
-  }
-
-  if (error) {
-    return <div>Error fetching data: {error.message}</div>;
-  }
-
-  return (
-    <PageLayout>
-      <div className="p-2">
-        {events.map((event) => (
-            <EventCard key={event.id} event={event} />
-          ))}
-      </div>
-    </PageLayout>
-  );
 }
 
 export default Home;
